requirements.txt
aiohttp
asyncio
discord-webhook
python-dotenv
